pyproject.toml
[project]
name = "mcp-nexus-rag"
version = "2.0.0"
description = "MCP Nexus Multi-Tenant RAG"
authors = [
    {name = "Turiya"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "mcp (>=1.26.0,<2.0.0)",
    "llama-index-core (>=0.14.15,<0.15.0)",
    "llama-index-llms-ollama (>=0.9.1,<0.10.0)",
    "llama-index-embeddings-ollama (>=0.8.6,<0.9.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "llama-index-postprocessor-flag-embedding-reranker (>=0.4.0,<0.5.0)",
    "FlagEmbedding (>=1.3.5,<2.0.0)",
    "transformers (>=4.40.0,<5.0.0)",
    "pathspec (>=1.0.4,<2.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.34.0,<1.0.0)",
    "watchdog (>=4.0.0,<5.0.0)",
    "llama-index-vector-stores-postgres (>=0.7.0,<0.8.0)",
    "llama-index-graph-stores-memgraph (>=0.4.0,<0.5.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "ruff (>=0.15.4,<0.16.0)",
    "pymarkdownlnt (>=0.9.35,<0.10.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-m 'not integration'"
markers = [
    "integration: tests that hit real databases (slow, run with: pytest -m integration)",
]

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true
